Script to add score column to bed files
Generated probability is interpreted as the annotation name, so we convert this to a score (integer, 0-1000) and append it to each row. Allows the markers to be shaded based on probability

+++ python/bedgraph_utils/add_score_column.py
@@ -17,7 +17,7 @@ def add_column(input, output, source_index=BED_COL_NAME, factor=1000.0):
     reader = csv.reader(input, delimiter='\t')
     writer = csv.writer(output, delimiter='\t')
     for row in reader:
-        row.append(float(row[BED_COL_NAME]) * factor)
+        row.append(int(float(row[BED_COL_NAME]) * factor))
         writer.writerow(row)
 
 if __name__ == '__main__':
